basic client-server construct

+++ blockchain.py
@@ -41,7 +41,7 @@ class Client:
                 time.sleep(5)
                 counter = 2
                 while True:
-                    proofResult = int(hex_dig) * counter
+                    proofResult = float(hex_dig) * counter
                     proofResult = str(proofResult)
                     print proofResult + "\n"
